Define ENR DB API ABC
